requirements.txt
celery==5.3.6
channels-redis==4.1.0
channels[daphne]==4.0.0
django-admin-rangefilter==0.12.0
django-elasticsearch-dsl-drf==0.22.5
django-elasticsearch-dsl==7.4
django-filter==23.4
django-model-utils==4.3.1
django-phonenumber-field[phonenumberslite]
Django==4.2.7
djangorestframework-bulk==0.2.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.14.0
drf-extra-fields==3.7.0
elasticsearch-dsl==7.4.1
elasticsearch==7.17.9
httpx==0.27.0
Pillow==10.1.0
psycopg2-binary==2.9.9
uvicorn[standard]==0.24.0.post1
moviepy==1.0.3
django-solo==2.3.0
django-admin-sortable2==2.2.1
